requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
